perf: switch to flat-playlist + skip-first + on-demand metadata fetch
Previous full-metadata mode wasted ~1 min/channel fetching data for
already-processed videos. New approach:

1. flat-playlist per tab → list video IDs quickly (~8s/channel)
2. Global skip check (glob output/*/**/*__videoID__*) before fetch
3. Only new videos trigger full metadata fetch + processing

Speed improvement: 3 channels with all videos processed:
  Before: ~3 min (full metadata for every video)
  After:  ~27s (flat listing + instant glob skip)

Also:
- Duration type back to float64 (flat-playlist returns float)
- Type filtering removed from FilterVideos (handled by tab suffix)
- IsProcessedGlobal: channel-agnostic skip detection

+++ fetcher/types.go
@@ -7,9 +7,9 @@ type VideoMeta struct {
 	Channel        string   `json:"uploader_id"`      // handle like @HighYield
 	ChannelName    string   `json:"channel"`           // display name like "High Yield"
 	UploadDate     string   `json:"upload_date"`       // YYYYMMDD
-	Duration       int      `json:"duration"`          // seconds
+	Duration       float64  `json:"duration"`          // seconds (float in flat-playlist, int in full)
 	DurationString string   `json:"duration_string"`
-	MediaType      string   `json:"media_type"`        // "short" for Shorts, empty for regular
+	MediaType      string   `json:"media_type"`        // "short" for Shorts (only in full metadata)
 	Language       string   `json:"language"`
 	Tags           []string `json:"tags"`
 	Categories     []string `json:"categories"`
